Reorganize font source
