+++ src/font/constants.js
@@ -1,0 +1,2 @@
+export const FONT_MSDF = 'msdf';
+export const FONT_BITMAP = 'bitmap';
